fixed minor errors and filenames
